pyproject.toml
[tool.poetry]
name = "line-message-forwarder"
version = "0.1.0"
description = ""
authors = ["CHANG-CHING-CHUNG <kangan987@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
line-bot-sdk = "^3.14.2"
flask = "^3.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
